refactor(spiral): move bucketing/aggregation/discsize into internal/spiral
