+++ internal/spiral/mode_category_test.go
@@ -1,4 +1,4 @@
-package main
+package spiral
 
 import (
 	"testing"
